python/requirements.txt
faster-whisper==1.1.1
nvidia-cublas-cu12==12.9.1.4
numpy==2.3.0
pynput==1.8.1
python-dotenv==1.2.2
requests==2.32.5
sounddevice==0.5.2
webrtcvad-wheels==2.0.14
